add location for properly exposing omnileads vue frontend
in particular, we add the `try_files`, required for a proper function of the
client-side router.

we should instead fix this using the same technique we use with the webui-*
locations, e.g. one for production(static) and another for development(rproxy)
but we are already on a deadline for OML-3004